+++ xbmc/filesystem/MythFile.h
@@ -51,7 +51,7 @@ public:
   virtual int64_t       GetLength();
   virtual int           Stat(const CURL& url, struct __stat64* buffer) { return -1; }
   virtual void          Close();
-  virtual unsigned int  Read(void* buffer, int64_t size);
+  virtual ssize_t       Read(void* buffer, size_t size);
   virtual std::string   GetContent() { return ""; }
   virtual bool          SkipNext();
